绕过腾讯审核发布CBA相关视频，以及其他视频类新闻
