+++ public/index.html
@@ -56,7 +56,7 @@
       // flexible 初始化
       flexible.init({
         designWidth: 750,
-        maxWidth: 414,
+        maxWidth: 750,
       });
       flexible.resize(function(info) {
         // console.log( info );
