+++ src/router/config.js
@@ -7,8 +7,9 @@ const pageRouterList = {
   sports: '/sports', // 体育新闻
   news: '/news', // 体育花边新闻
   photos: '/photos', // 图集新闻
-  videos: '/videos', // CBA优酷视频
-  cbaVideos: '/cbaVideos', // CBA优酷视频列表页
+  videos: '/videos', // CBA优酷视频(嵌套iframe)
+  cbaNews: '/cbaNews', // CBA新闻页
+  cbaVideos: '/cbaVideos', // CBA优酷视频页（直接跳转）
 };
 export default pageRouterList;
